+++ tests/form/lightFilter.test.tsx
@@ -11,6 +11,7 @@ import {
 } from '@ant-design/pro-form';
 import {
   act,
+  cleanup,
   fireEvent,
   render,
   screen,
@@ -19,6 +20,10 @@ import {
 import userEvent from '@testing-library/user-event';
 import dayjs from 'dayjs';
 import KeyCode from 'rc-util/es/KeyCode';
+
+afterEach(() => {
+  cleanup();
+});
 
 describe('LightFilter', () => {
   it(' 🪕 basic use text', async () => {
@@ -762,7 +767,7 @@ describe('LightFilter', () => {
 
     await waitFor(() => {
       expect(
-        container.querySelector('.ant-pro-core-field-label')?.textContent,
+        container.querySelector('.ant-pro-core-field-label'),
       ).toMatchSnapshot();
     });
   });