+++ tests/table/dynamic-columns-state.test.tsx
@@ -1,5 +1,5 @@
 import { ProTable } from '@ant-design/pro-components';
-import { act, render } from '@testing-library/react';
+import { act, cleanup, render } from '@testing-library/react';
 import { waitForWaitTime } from '../util';
 
 const valueEnum = {
@@ -40,6 +40,10 @@ for (let i = 0; i < 5; i += 1) {
     statusText: '这是一段很随意的文字',
   });
 }
+
+afterEach(() => {
+  cleanup();
+});
 
 describe('Dynamic Persistence', () => {
   it('🎏 columnSetting columnsState.persistenceKey change', async () => {