+++ tests/table/selectKeys.test.tsx
@@ -1,8 +1,12 @@
 import ProTable from '@ant-design/pro-table';
-import { act, render, waitFor } from '@testing-library/react';
+import { act, cleanup, render, waitFor } from '@testing-library/react';
 import React, { useState } from 'react';
 import { waitForWaitTime } from '../util';
 import { getFetchData } from './demo';
+
+afterEach(() => {
+  cleanup();
+});
 
 describe('BasicTable Search', () => {
   const LINE_STR_COUNT = 20;