+++ tests/table/valueEnum.test.tsx
@@ -1,6 +1,6 @@
 import ProProvider from '@ant-design/pro-provider';
 import ProTable from '@ant-design/pro-table';
-import { act, render, waitFor } from '@testing-library/react';
+import { act, cleanup, render, waitFor } from '@testing-library/react';
 import { Input } from 'antd';
 import { useContext } from 'react';
 import { waitForWaitTime } from '../util';
@@ -46,6 +46,10 @@ const Demo = () => {
     </ProProvider.Provider>
   );
 };
+
+afterEach(() => {
+  cleanup();
+});
 
 describe('Table valueEnum', () => {
   it('🎏 dynamic enum test', async () => {